+++ docs/history.rst
@@ -5,6 +5,9 @@ Release history
 ---------------
 - Add :data:`~klempner.config.URL_SCHEME_MAP` so that URL schemes can be
   easily configured.
+- Separate the :ref:`environment-discovery-method` method out from the
+  :ref:`simple-discovery-method` method.  Previously the simple method would
+  fall back to using environment variables if they were set.
 
 `0.0.1`_ (5 May 2019)
 ---------------------